Add FunctionParameterToolParser and refactor existing parsers
- Introduced `FunctionParameterToolParser` to handle tool calls in the `<function=...><parameter=...>` format, enhancing parsing capabilities for models like Qwen3Coder and Nemotron3Nano.
- Updated `Nemotron3NanoToolParser` and `Qwen3CoderToolParser` to inherit from `FunctionParameterToolParser`, streamlining their implementations.
- Refactored `extract_tool_calls_streaming` method in `AbstractToolParser` to improve tool call extraction from streaming chunks.
- Removed redundant streaming extraction methods from `GLM4MoEToolParser`, `HermesToolParser`, and other parsers, consolidating functionality.
