+++ app/parsers/functiongemma.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import re
 import json
 
-from .abstract_parser import AbstractToolParser, ToolParserState
+from .abstract_parser import AbstractToolParser
 
 TOOL_OPEN = "<start_function_call>"
 TOOL_CLOSE = "<end_function_call>"
@@ -60,48 +60,3 @@ class FunctionGemmaToolParser(AbstractToolParser):
             "tool_calls": tool_calls,
         }
         
-
-    def extract_tool_calls_streaming(
-        self, chunk: str
-    ) -> tuple[dict[str, list] | str | None, bool]:
-        """Extract tool calls from streaming chunks.
-        
-        Parameters
-        ----------
-        chunk : str
-            Chunk of model output to process.
-            
-        Returns
-        -------
-        tuple[dict[str, list] | str | None, bool]
-            Tuple of (extracted_content, is_complete) where:
-            - extracted_content: Tool calls dict if complete, chunk if passthrough, None if buffering
-            - is_complete: True if chunk should be sent, False if still buffering
-        """
-        if self.tool_open in chunk:
-            self.state = ToolParserState.FOUND_PREFIX
-
-            if self.tool_close in chunk:
-                self.buffer += chunk
-                result = self.extract_tool_calls(self.buffer)
-                self.buffer = ""
-                self.state = ToolParserState.NORMAL
-                return result, True
-            else:
-                self.buffer += chunk
-                return None, False
-
-        if self.state == ToolParserState.FOUND_PREFIX:
-            if self.tool_close in chunk:
-                self.buffer += chunk
-                result = self.extract_tool_calls(self.buffer)
-                self.buffer = ""
-                self.state = ToolParserState.NORMAL
-                return result, True
-            else:
-                self.buffer += chunk
-                return None, False
-
-        return {
-            "content": chunk
-        }, False
